Fix/design task 01

+++ lib/layouts/topics_talk.dart
@@ -75,15 +75,17 @@ class _TopicsTalk extends State<TopicsTalk> {
                   setState(() {
                     loading = true;
                   });
+                  await userService.deleteTopicsHear(userId);
+                  await userService.setTopicsHear(userId, topicsTalk);
 
-                  if (changed) {
+                  if (changed) {         
+                    if (deletedTopics.length > 0) {
+                      await userService.deleteTopicsTalk(userId, deletedTopics);
+                    }
                     userService.setTopicsTalk(userId, topicsTalk);
-                    userService.deleteTopicsTalk(userId, deletedTopics);
-
-                    await userService.deleteTopicsHear(userId);
-                    await userService.setTopicsHear(userId, topicsTalk);
                     deletedTopics.clear();
                   }
+
                   List<String> segments = await userService.getSegments(userId);
 
                   await Navigator.of(context)
